RavenDB-13110 - Incoming Replication Handler abstraction
