+++ src/Raven.Server/Documents/Replication/Incoming/AbstractIncomingReplicationHandler.cs
@@ -1,72 +1,84 @@
 ﻿using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
+using System.Linq;
 using System.Net;
 using System.Net.Sockets;
 using System.Threading;
+using Raven.Client.Documents.Replication;
 using Raven.Client.Documents.Replication.Messages;
+using Raven.Client.Extensions;
 using Raven.Client.ServerWide.Tcp;
+using Raven.Server.Documents.Replication.Outgoing;
 using Raven.Server.Documents.Replication.ReplicationItems;
 using Raven.Server.Documents.Replication.Stats;
 using Raven.Server.Documents.TcpHandlers;
+using Raven.Server.Exceptions;
 using Raven.Server.ServerWide;
 using Raven.Server.Utils;
-using Sparrow;
 using Sparrow.Json;
+using Sparrow.Json.Parsing;
+using Sparrow.Json.Sync;
 using Sparrow.Logging;
-using Sparrow.Platform;
 using Sparrow.Server;
 using Sparrow.Threading;
 using Sparrow.Utils;
+using Voron;
 
 namespace Raven.Server.Documents.Replication.Incoming
 {
-    public abstract class AbstractIncomingReplicationHandler : IDisposable
+    public abstract class AbstractIncomingReplicationHandler<TContextPool, TOperationContext> : IDisposable
+    where TContextPool : JsonContextPoolBase<TOperationContext>
+    where TOperationContext : JsonOperationContext
     {
-        protected PoolOfThreads.LongRunningWork _incomingWork;
-        protected readonly DisposeOnce<SingleAttempt> _disposeOnce;
-        protected Logger _log;
-        protected readonly TcpClient _tcpClient;
+        private PoolOfThreads.LongRunningWork _incomingWork;
+        private readonly DisposeOnce<SingleAttempt> _disposeOnce;
+        private readonly ServerStore _server;
+        private IDisposable _connectionOptionsDisposable;
+        private readonly TContextPool _contextPool;
+        private readonly string _databaseName;
+        private readonly ConcurrentQueue<IncomingReplicationStatsAggregator> _lastReplicationStats = new ConcurrentQueue<IncomingReplicationStatsAggregator>();
+        private IncomingReplicationStatsAggregator _lastStats;
+        private readonly TcpClient _tcpClient;
+
         protected readonly Stream _stream;
         protected readonly TcpConnectionOptions _tcpConnectionOptions;
-        private readonly ServerStore _server;
         protected readonly (IDisposable ReleaseBuffer, JsonOperationContext.MemoryBuffer Buffer) _copiedBuffer;
         protected readonly CancellationTokenSource _cts;
         protected StreamsTempFile _attachmentStreamsTempFile;
+        protected long _lastDocumentEtag;
+        protected AsyncManualResetEvent _replicationFromAnotherSource;
 
-        public ServerStore Server => _server;
-        public string DatabaseName { get; }
+        protected Logger Logger;
 
         public IncomingConnectionInfo ConnectionInfo { get; protected set; }
-
         public TcpConnectionHeaderMessage.SupportedFeatures SupportedFeatures { get; set; }
-
         protected string IncomingReplicationThreadName => $"Incoming replication {FromToString}";
-
-        public virtual string FromToString => $"In database {Server.NodeTag}-{DatabaseName} @ {Server.GetNodeTcpServerUrl()} " +
+        public virtual string FromToString => $"In database {_server.NodeTag}-{_databaseName} @ {_server.GetNodeTcpServerUrl()} " +
                                               $"from {ConnectionInfo.SourceTag}-{ConnectionInfo.SourceDatabaseName} @ {ConnectionInfo.SourceUrl}";
 
-        public bool IsDisposed => _disposeOnce.Disposed;
-
-        protected AbstractIncomingReplicationHandler(TcpConnectionOptions tcpConnectionOptions, JsonOperationContext.MemoryBuffer buffer, ServerStore server, string databaseName, ReplicationLatestEtagRequest replicatedLastEtag, CancellationToken token)
+        protected AbstractIncomingReplicationHandler(TcpConnectionOptions tcpConnectionOptions, JsonOperationContext.MemoryBuffer buffer,
+            ServerStore server, string databaseName, ReplicationLatestEtagRequest replicatedLastEtag, CancellationToken token, TContextPool contextPool)
         {
             _disposeOnce = new DisposeOnce<SingleAttempt>(DisposeInternal);
             _tcpConnectionOptions = tcpConnectionOptions;
             _copiedBuffer = buffer.Clone(_tcpConnectionOptions.ContextPool);
-            _log = LoggingSource.Instance.GetLogger(databaseName, GetType().FullName);
+
             _tcpClient = tcpConnectionOptions.TcpClient;
             _stream = tcpConnectionOptions.Stream;
             _server = server;
             _cts = CancellationTokenSource.CreateLinkedTokenSource(token);
+            _databaseName = databaseName;
+            _contextPool = contextPool;
 
-            DatabaseName = databaseName;
+            Logger = LoggingSource.Instance.GetLogger(databaseName, GetType().FullName);
+
             ConnectionInfo = IncomingConnectionInfo.FromGetLatestEtag(replicatedLastEtag);
             SupportedFeatures = TcpConnectionHeaderMessage.GetSupportedFeaturesFor(tcpConnectionOptions.Operation, tcpConnectionOptions.ProtocolVersion);
             ConnectionInfo.RemoteIp = ((IPEndPoint)_tcpClient.Client.RemoteEndPoint)?.Address.ToString();
         }
-
-        protected abstract void ReceiveReplicationBatches();
 
         public void Start()
         {
@@ -81,8 +93,8 @@ namespace Raven.Server.Documents.Replication.Incoming
                 _incomingWork = PoolOfThreads.GlobalRavenThreadPool.LongRunning(x => { DoIncomingReplication(); }, null, IncomingReplicationThreadName);
             }
 
-            if (_log.IsInfoEnabled)
-                _log.Info($"Incoming replication thread started ({FromToString})");
+            if (Logger.IsInfoEnabled)
+                Logger.Info($"Incoming replication thread started ({FromToString})");
         }
 
         public void DoIncomingReplication()
@@ -93,22 +105,326 @@ namespace Raven.Server.Documents.Replication.Incoming
             }
             catch (Exception e)
             {
-                if (_log.IsInfoEnabled)
-                    _log.Info($"Error in accepting replication request ({FromToString})", e);
+                if (Logger.IsInfoEnabled)
+                    Logger.Info($"Error in accepting replication request ({FromToString})", e);
             }
         }
 
-        protected ReplicationBatchItem ReadItemFromSource(Reader reader, JsonOperationContext context, ByteStringContext allocator, IncomingReplicationStatsScope stats)
+        protected void ReceiveReplicationBatches()
         {
-            stats.RecordInputAttempt();
+            NativeMemory.EnsureRegistered();
+            try
+            {
+                using (_connectionOptionsDisposable = _tcpConnectionOptions.ConnectionProcessingInProgress("Replication"))
+                using (_stream)
+                using (var interruptibleRead = new InterruptibleRead<TContextPool, TOperationContext>(_contextPool, _stream))
+                {
+                    while (_cts.IsCancellationRequested == false)
+                    {
+                        try
+                        {
+                            AddReplicationPulse(ReplicationPulseDirection.IncomingInitiate);
 
-            var item = ReplicationBatchItem.ReadTypeAndInstantiate(reader);
-            item.ReadChangeVectorAndMarker();
-            item.Read(context, allocator, stats);
+                            using (var msg = interruptibleRead.ParseToMemory(
+                                _replicationFromAnotherSource,
+                                "IncomingReplication/read-message",
+                                Timeout.Infinite,
+                                _copiedBuffer.Buffer,
+                                _cts.Token))
+                            {
+                                if (msg.Document != null)
+                                {
+                                    EnsureNotDeleted(_server.NodeTag);
 
-            return item;
+                                    using (var writer = new BlittableJsonTextWriter(msg.Context, _stream))
+                                    {
+                                        HandleSingleReplicationBatch(msg.Context,
+                                            msg.Document,
+                                            writer);
+                                    }
+                                }
+                                else // notify peer about new change vector
+                                {
+                                    using (_contextPool.AllocateOperationContext(out TOperationContext context))
+                                    using (var writer = new BlittableJsonTextWriter(context, _stream))
+                                    {
+                                        SendHeartbeatStatusToSource(
+                                            context,
+                                            writer,
+                                            _lastDocumentEtag,
+                                            "Notify");
+                                    }
+                                }
+                                // we reset it after every time we send to the remote server
+                                // because that is when we know that it is up to date with our
+                                // status, so no need to send again
+                                _replicationFromAnotherSource?.Reset();
+                            }
+                        }
+                        catch (Exception e)
+                        {
+                            AddReplicationPulse(ReplicationPulseDirection.IncomingInitiateError, e.Message);
+
+                            if (Logger.IsInfoEnabled)
+                            {
+                                if (e is AggregateException ae &&
+                                    ae.InnerExceptions.Count == 1 &&
+                                    ae.InnerException is SocketException ase)
+                                {
+                                    HandleSocketException(ase);
+                                }
+                                else if (e.InnerException is SocketException se)
+                                {
+                                    HandleSocketException(se);
+                                }
+                                else
+                                {
+                                    //if we are disposing, do not notify about failure (not relevant)
+                                    if (_cts.IsCancellationRequested == false)
+                                        if (Logger.IsInfoEnabled)
+                                            Logger.Info("Received unexpected exception while receiving replication batch.", e);
+                                }
+                            }
+
+                            throw;
+                        }
+
+                        void HandleSocketException(SocketException e)
+                        {
+                            if (Logger.IsInfoEnabled)
+                                Logger.Info("Failed to read data from incoming connection. The incoming connection will be closed and re-created.", e);
+                        }
+                    }
+                }
+            }
+            catch (Exception e)
+            {
+                //if we are disposing, do not notify about failure (not relevant)
+                if (_cts.IsCancellationRequested == false)
+                {
+                    if (Logger.IsInfoEnabled)
+                        Logger.Info($"Connection error {FromToString}: an exception was thrown during receiving incoming document replication batch.", e);
+
+                    OnFailed(e);
+                }
+            }
         }
 
+        internal void HandleSingleReplicationBatch(
+    TOperationContext context,
+    BlittableJsonReaderObject message,
+    BlittableJsonTextWriter writer)
+        {
+            message.BlittableValidation();
+            //note: at this point, the valid messages are heartbeat and replication batch.
+            _cts.Token.ThrowIfCancellationRequested();
+            string messageType = null;
+            try
+            {
+                if (!message.TryGet(nameof(ReplicationMessageHeader.Type), out messageType))
+                    throw new InvalidDataException("Expected the message to have a 'Type' field. The property was not found");
+
+                if (!message.TryGet(nameof(ReplicationMessageHeader.LastDocumentEtag), out _lastDocumentEtag))
+                    throw new InvalidOperationException("Expected LastDocumentEtag property in the replication message, " +
+                                                        "but didn't find it..");
+
+                switch (messageType)
+                {
+                    case ReplicationMessageType.Documents:
+                        AddReplicationPulse(ReplicationPulseDirection.IncomingBegin);
+
+                        var stats = _lastStats = new IncomingReplicationStatsAggregator(GetNextReplicationStatsId(), _lastStats);
+                        AddReplicationPerformance(stats);
+
+                        try
+                        {
+                            using (var scope = stats.CreateScope())
+                            {
+                                try
+                                {
+                                    scope.RecordLastEtag(_lastDocumentEtag);
+
+                                    HandleReceivedDocumentsAndAttachmentsBatch(context, message, _lastDocumentEtag, scope);
+                                    break;
+                                }
+                                catch (Exception e)
+                                {
+                                    AddReplicationPulse(ReplicationPulseDirection.IncomingError, e.Message);
+                                    scope.AddError(e);
+                                    throw;
+                                }
+                            }
+                        }
+                        finally
+                        {
+                            AddReplicationPulse(ReplicationPulseDirection.IncomingEnd);
+                            stats.Complete();
+                        }
+                    case ReplicationMessageType.Heartbeat:
+                        AddReplicationPulse(ReplicationPulseDirection.IncomingHeartbeat);
+
+                        HandleHeartbeatMessage(context, message);
+                        break;
+
+                    default:
+                        throw new ArgumentOutOfRangeException("Unknown message type: " + messageType);
+                }
+
+                SendHeartbeatStatusToSource(context, writer, _lastDocumentEtag, messageType);
+            }
+            catch (ObjectDisposedException)
+            {
+                //we are shutting down replication, this is ok
+            }
+            catch (EndOfStreamException e)
+            {
+                if (Logger.IsInfoEnabled)
+                    Logger.Info("Received unexpected end of stream while receiving replication batches. " +
+                              "This might indicate an issue with network.", e);
+                throw;
+            }
+            catch (Exception e)
+            {
+                //if we are disposing, ignore errors
+                if (_cts.IsCancellationRequested)
+                    return;
+
+                DynamicJsonValue returnValue;
+
+                if (e.ExtractSingleInnerException() is MissingAttachmentException mae)
+                {
+                    returnValue = new DynamicJsonValue
+                    {
+                        [nameof(ReplicationMessageReply.Type)] = ReplicationMessageReply.ReplyType.MissingAttachments.ToString(),
+                        [nameof(ReplicationMessageReply.MessageType)] = messageType,
+                        [nameof(ReplicationMessageReply.LastEtagAccepted)] = -1,
+                        [nameof(ReplicationMessageReply.Exception)] = mae.ToString()
+                    };
+
+                    context.Write(writer, returnValue);
+                    writer.Flush();
+
+                    return;
+                }
+
+                if (Logger.IsInfoEnabled)
+                    Logger.Info($"Failed replicating documents {FromToString}.", e);
+
+                //return negative ack
+                returnValue = new DynamicJsonValue
+                {
+                    [nameof(ReplicationMessageReply.Type)] = ReplicationMessageReply.ReplyType.Error.ToString(),
+                    [nameof(ReplicationMessageReply.MessageType)] = messageType,
+                    [nameof(ReplicationMessageReply.LastEtagAccepted)] = -1,
+                    [nameof(ReplicationMessageReply.Exception)] = e.ToString()
+                };
+
+                context.Write(writer, returnValue);
+                writer.Flush();
+
+                throw;
+            }
+        }
+
+        protected void AddReplicationPerformance(IncomingReplicationStatsAggregator stats)
+        {
+            _lastReplicationStats.Enqueue(stats);
+
+            while (_lastReplicationStats.Count > 25)
+                _lastReplicationStats.TryDequeue(out stats);
+        }
+
+        public IncomingReplicationPerformanceStats[] GetReplicationPerformance()
+        {
+            var lastStats = _lastStats;
+
+            return _lastReplicationStats
+                .Select(x => x == lastStats ? x.ToReplicationPerformanceLiveStatsWithDetails() : x.ToReplicationPerformanceStats())
+                .ToArray();
+        }
+
+        public IncomingReplicationStatsAggregator GetLatestReplicationPerformance()
+        {
+            return _lastStats;
+        }
+
+        private void HandleReceivedDocumentsAndAttachmentsBatch(TOperationContext context, BlittableJsonReaderObject message, long lastDocumentEtag, IncomingReplicationStatsScope stats)
+        {
+            if (!message.TryGet(nameof(ReplicationMessageHeader.ItemsCount), out int itemsCount))
+                throw new InvalidDataException($"Expected the '{nameof(ReplicationMessageHeader.ItemsCount)}' field, " +
+                                               $"but had no numeric field of this value, this is likely a bug");
+
+            if (!message.TryGet(nameof(ReplicationMessageHeader.AttachmentStreamsCount), out int attachmentStreamCount))
+                throw new InvalidDataException($"Expected the '{nameof(ReplicationMessageHeader.AttachmentStreamsCount)}' field, " +
+                                               $"but had no numeric field of this value, this is likely a bug");
+
+            ReceiveSingleDocumentsBatch(context, itemsCount, attachmentStreamCount, lastDocumentEtag, stats);
+
+            OnAttachmentStreamsReceives(attachmentStreamCount);
+
+            OnDocumentsReceived();
+        }
+
+        protected void ReadItemsFromSource(int replicatedDocs, TOperationContext context, ByteStringContext allocator, IncomingReplicationHandler.DataForReplicationCommand data, Reader reader,
+            IncomingReplicationStatsScope stats)
+        {
+            if (data.ReplicatedItems == null)
+                data.ReplicatedItems = new ReplicationBatchItem[replicatedDocs];
+
+            for (var i = 0; i < replicatedDocs; i++)
+            {
+                stats.RecordInputAttempt();
+
+                var item = ReplicationBatchItem.ReadTypeAndInstantiate(reader);
+                item.ReadChangeVectorAndMarker();
+                item.Read(context, allocator, stats);
+
+                data.ReplicatedItems[i] = item;
+            }
+        }
+
+        protected void ReadAttachmentStreamsFromSource(int attachmentStreamCount,
+            TOperationContext context, ByteStringContext allocator, IncomingReplicationHandler.DataForReplicationCommand dataForReplicationCommand, Reader reader, IncomingReplicationStatsScope stats)
+        {
+            if (attachmentStreamCount == 0)
+                return;
+
+            var replicatedAttachmentStreams = new Dictionary<Slice, AttachmentReplicationItem>(SliceComparer.Instance);
+
+            for (var i = 0; i < attachmentStreamCount; i++)
+            {
+                var attachment = (AttachmentReplicationItem)ReplicationBatchItem.ReadTypeAndInstantiate(reader);
+                Debug.Assert(attachment.Type == ReplicationBatchItem.ReplicationItemType.AttachmentStream);
+
+                using (stats.For(ReplicationOperation.Incoming.AttachmentRead))
+                {
+                    attachment.ReadStream(allocator, _attachmentStreamsTempFile);
+                    replicatedAttachmentStreams[attachment.Base64Hash] = attachment;
+                }
+            }
+
+            dataForReplicationCommand.ReplicatedAttachmentStreams = replicatedAttachmentStreams;
+        }
+
+        protected abstract void EnsureNotDeleted(string nodeTag);
+
+        protected abstract void OnDocumentsReceived();
+
+        protected abstract void OnAttachmentStreamsReceives(int attachmentStreamCount);
+
+        protected abstract void OnFailed(Exception exception);
+
+        protected abstract void ReceiveSingleDocumentsBatch(TOperationContext context, int itemsCount, int attachmentStreamCount, long lastDocumentEtag, IncomingReplicationStatsScope stats);
+
+        protected abstract int GetNextReplicationStatsId();
+
+        internal abstract void SendHeartbeatStatusToSource(TOperationContext context, BlittableJsonTextWriter writer, long lastDocumentEtag, string notify);
+
+        protected abstract void AddReplicationPulse(ReplicationPulseDirection direction, string exceptionMessage = null);
+
+        protected abstract void HandleHeartbeatMessage(TOperationContext jsonOperationContext, BlittableJsonReaderObject blittableJsonReaderObject);
+
+        public bool IsDisposed => _disposeOnce.Disposed;
 
         public void Dispose()
         {
@@ -120,10 +436,17 @@ namespace Raven.Server.Documents.Replication.Incoming
             var releaser = _copiedBuffer.ReleaseBuffer;
             try
             {
-                if (_log.IsInfoEnabled)
-                    _log.Info($"Disposing IncomingReplicationHandler ({FromToString})");
+                if (Logger.IsInfoEnabled)
+                    Logger.Info($"Disposing IncomingReplicationHandler ({FromToString})");
                 _cts.Cancel();
 
+                try
+                {
+                    _connectionOptionsDisposable?.Dispose();
+                }
+                catch (Exception)
+                {
+                }
                 try
                 {
                     _stream.Dispose();
@@ -164,6 +487,7 @@ namespace Raven.Server.Documents.Replication.Incoming
                 _cts.Dispose();
 
                 _attachmentStreamsTempFile.Dispose();
+                _replicationFromAnotherSource?.Dispose();
             }
             finally
             {
@@ -174,106 +498,6 @@ namespace Raven.Server.Documents.Replication.Incoming
                 catch (Exception)
                 {
                     // can't do anything about it...
-                }
-            }
-        }
-
-        public unsafe class IncomingReplicationAllocator : IDisposable
-        {
-            private readonly long _maxSizeForContextUseInBytes;
-            private readonly long _minSizeToAllocateNonContextUseInBytes;
-            public long TotalDocumentsSizeInBytes { get; private set; }
-
-            private List<Allocation> _nativeAllocationList;
-            private Allocation _currentAllocation;
-
-            protected ByteStringContext _allocator;
-
-            public IncomingReplicationAllocator(ByteStringContext allocator, Size? maxSizeToSend)
-            {
-                _allocator = allocator;
-                var maxSizeForContextUse = maxSizeToSend * 2 ?? new Size(128, SizeUnit.Megabytes);
-
-                _maxSizeForContextUseInBytes = maxSizeForContextUse.GetValue(SizeUnit.Bytes);
-                var minSizeToNonContextAllocationInMb = PlatformDetails.Is32Bits ? 4 : 16;
-                _minSizeToAllocateNonContextUseInBytes = new Size(minSizeToNonContextAllocationInMb, SizeUnit.Megabytes).GetValue(SizeUnit.Bytes);
-            }
-
-            public byte* AllocateMemory(int size)
-            {
-                TotalDocumentsSizeInBytes += size;
-                if (TotalDocumentsSizeInBytes <= _maxSizeForContextUseInBytes)
-                {
-                    _allocator.Allocate(size, out var output);
-                    return output.Ptr;
-                }
-
-                if (_currentAllocation == null || _currentAllocation.Free < size)
-                {
-                    // first allocation or we don't have enough space on the currently allocated chunk
-
-                    // there can be a document that is larger than the minimum
-                    var sizeToAllocate = Math.Max(size, _minSizeToAllocateNonContextUseInBytes);
-
-                    var allocation = new Allocation(sizeToAllocate);
-                    if (_nativeAllocationList == null)
-                        _nativeAllocationList = new List<Allocation>();
-
-                    _nativeAllocationList.Add(allocation);
-                    _currentAllocation = allocation;
-                }
-
-                return _currentAllocation.GetMemory(size);
-            }
-
-            public void Dispose()
-            {
-                if (_nativeAllocationList == null)
-                    return;
-
-                foreach (var allocation in _nativeAllocationList)
-                {
-                    allocation.Dispose();
-                }
-            }
-
-            private class Allocation : IDisposable
-            {
-                private readonly byte* _ptr;
-                private readonly long _allocationSize;
-                private readonly NativeMemory.ThreadStats _threadStats;
-                private long _used;
-                public long Free => _allocationSize - _used;
-
-                public Allocation(long allocationSize)
-                {
-                    _ptr = NativeMemory.AllocateMemory(allocationSize, out var threadStats);
-                    _allocationSize = allocationSize;
-                    _threadStats = threadStats;
-                }
-
-                public byte* GetMemory(long size)
-                {
-                    ThrowOnPointerOutOfRange(size);
-
-                    var mem = _ptr + _used;
-                    _used += size;
-                    return mem;
-                }
-
-                [Conditional("DEBUG")]
-                private void ThrowOnPointerOutOfRange(long size)
-                {
-                    if (_used + size > _allocationSize)
-                        throw new InvalidOperationException(
-                            $"Not enough space to allocate the requested size: {new Size(size, SizeUnit.Bytes)}, " +
-                            $"used: {new Size(_used, SizeUnit.Bytes)}, " +
-                            $"total allocation size: {new Size(_allocationSize, SizeUnit.Bytes)}");
-                }
-
-                public void Dispose()
-                {
-                    NativeMemory.Free(_ptr, _allocationSize, _threadStats);
                 }
             }
         }
