+++ src/Raven.Server/Documents/Sharding/Handlers/ShardedIncomingReplicationHandler.cs
@@ -2,14 +2,12 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using System.Net.Sockets;
 using System.Threading;
 using Raven.Client.Documents.Replication.Messages;
 using Raven.Client.Exceptions.Sharding;
 using Raven.Client.Extensions;
 using Raven.Server.Documents.Replication;
 using Raven.Server.Documents.Replication.Incoming;
-using Raven.Server.Documents.Replication.Outgoing;
 using Raven.Server.Documents.Replication.ReplicationItems;
 using Raven.Server.Documents.Replication.Stats;
 using Raven.Server.Documents.TcpHandlers;
@@ -18,241 +16,85 @@ using Raven.Server.ServerWide.Context;
 using Sparrow.Json;
 using Sparrow.Json.Parsing;
 using Sparrow.Json.Sync;
-using Sparrow.Server.Json.Sync;
-using Sparrow.Utils;
 using Voron;
 
 namespace Raven.Server.Documents.Sharding.Handlers
 {
-    public class ShardedIncomingReplicationHandler : AbstractIncomingReplicationHandler
+    public class ShardedIncomingReplicationHandler : AbstractIncomingReplicationHandler<TransactionContextPool, TransactionOperationContext>
     {
         private readonly ShardedDatabaseContext.ShardedReplicationContext _parent;
         private readonly ReplicationQueue _replicationQueue;
-        private long _lastDocumentEtag;
-
-        public long LastDocumentEtag => _lastDocumentEtag;
 
         public ShardedIncomingReplicationHandler(TcpConnectionOptions tcpConnectionOptions, ShardedDatabaseContext.ShardedReplicationContext parent,
             JsonOperationContext.MemoryBuffer buffer, ReplicationLatestEtagRequest replicatedLastEtag, ReplicationQueue replicationQueue)
-            : base(tcpConnectionOptions, buffer, parent.Server, parent.DatabaseName, replicatedLastEtag, parent.Context.DatabaseShutdown)
+            : base(tcpConnectionOptions, buffer, parent.Server, parent.DatabaseName, replicatedLastEtag, parent.Context.DatabaseShutdown, parent.Server.ContextPool)
         {
             _parent = parent;
             _replicationQueue = replicationQueue;
             _attachmentStreamsTempFile = new StreamsTempFile("ShardedReplication" + Guid.NewGuid(), false);
         }
 
-        protected override void ReceiveReplicationBatches()
+        protected override void EnsureNotDeleted(string nodeTag) => _parent.EnsureNotDeleted(_parent.Server.NodeTag);
+
+        protected override void OnFailed(Exception exception)
         {
-            NativeMemory.EnsureRegistered();
-            try
-            {
-
-                using (_tcpConnectionOptions.ConnectionProcessingInProgress("Replication"))
-                using (_parent.Server.ContextPool.AllocateOperationContext(out TransactionOperationContext context))
-                using (_stream)
-                {
-                    while (_cts.IsCancellationRequested == false)
-                    {
-                        try
-                        {
-                            using (var msg = context.Sync.ParseToMemory(
-                                       _stream,
-                                       "IncomingReplication/read-message",
-                                       BlittableJsonDocumentBuilder.UsageMode.None,
-                                       _copiedBuffer.Buffer))
-                            {
-                                if (msg != null)
-                                {
-                                    _parent.EnsureNotDeleted(_parent.Server.NodeTag);
-
-                                    using (var writer = new BlittableJsonTextWriter(context, _stream))
-                                    {
-                                        HandleSingleReplicationBatch(context,
-                                            msg,
-                                            writer);
-                                    }
-                                }
-                                else // notify peer about new change vector
-                                {
-                                    using (var writer = new BlittableJsonTextWriter(context, _stream))
-                                    {
-                                        SendHeartbeatStatusToSource(
-                                            context,
-                                            writer,
-                                            _lastDocumentEtag,
-                                            "Notify");
-                                    }
-                                }
-                            }
-                        }
-                        catch (Exception e)
-                        {
-                            if (_log.IsInfoEnabled)
-                            {
-                                if (e is AggregateException ae &&
-                                    ae.InnerExceptions.Count == 1 &&
-                                    ae.InnerException is SocketException ase)
-                                {
-                                    HandleSocketException(ase);
-                                }
-                                else if (e.InnerException is SocketException se)
-                                {
-                                    HandleSocketException(se);
-                                }
-                                else
-                                {
-                                    //if we are disposing, do not notify about failure (not relevant)
-                                    if (_parent.Context.DatabaseShutdown.IsCancellationRequested == false)
-                                        if (_log.IsInfoEnabled)
-                                            _log.Info("Received unexpected exception while receiving replication batch.", e);
-                                }
-                            }
-
-                            throw;
-                        }
-
-                        void HandleSocketException(SocketException e)
-                        {
-                            if (_log.IsInfoEnabled)
-                                _log.Info("Failed to read data from incoming connection. The incoming connection will be closed and re-created.", e);
-                        }
-                    }
-                }
-            }
-            catch (Exception e)
-            {
-                //if we are disposing, do not notify about failure (not relevant)
-                if (_cts.IsCancellationRequested == false)
-                {
-                    if (_log.IsInfoEnabled)
-                        _log.Info($"Connection error {FromToString}: an exception was thrown during receiving incoming document replication batch.", e);
-                }
-            }
         }
 
-        private void HandleSingleReplicationBatch(
-    TransactionOperationContext context,
-    BlittableJsonReaderObject message,
-    BlittableJsonTextWriter writer)
+        protected override void HandleHeartbeatMessage(TransactionOperationContext jsonOperationContext, BlittableJsonReaderObject blittableJsonReaderObject)
         {
-            message.BlittableValidation();
-            //note: at this point, the valid messages are heartbeat and replication batch.
-            _cts.Token.ThrowIfCancellationRequested();
-            string messageType = null;
-            try
-            {
-                if (!message.TryGet(nameof(ReplicationMessageHeader.Type), out messageType))
-                    throw new InvalidDataException("Expected the message to have a 'Type' field. The property was not found");
-
-                if (!message.TryGet(nameof(ReplicationMessageHeader.LastDocumentEtag), out _lastDocumentEtag))
-                    throw new InvalidOperationException("Expected LastDocumentEtag property in the replication message, " +
-                                                        "but didn't find it..");
-
-                switch (messageType)
-                {
-                    case ReplicationMessageType.Documents:
-
-                        if (!message.TryGet(nameof(ReplicationMessageHeader.ItemsCount), out int itemsCount))
-                            throw new InvalidDataException($"Expected the '{nameof(ReplicationMessageHeader.ItemsCount)}' field, " +
-                                                           $"but had no numeric field of this value, this is likely a bug");
-
-                        if (!message.TryGet(nameof(ReplicationMessageHeader.AttachmentStreamsCount), out int attachmentStreamCount))
-                            throw new InvalidDataException($"Expected the '{nameof(ReplicationMessageHeader.AttachmentStreamsCount)}' field, " +
-                                                           $"but had no numeric field of this value, this is likely a bug");
-
-                        ReceiveSingleDocumentsBatch(context, itemsCount, attachmentStreamCount);
-
-                        break;
-
-                    case ReplicationMessageType.Heartbeat:
-                        break;
-
-                    default:
-                        throw new ArgumentOutOfRangeException("Unknown message type: " + messageType);
-                }
-
-                SendHeartbeatStatusToSource(context, writer, _lastDocumentEtag, messageType);
-
-            }
-            catch (ObjectDisposedException)
-            {
-                //we are shutting down replication, this is ok
-            }
-            catch (EndOfStreamException e)
-            {
-                if (_log.IsInfoEnabled)
-                    _log.Info("Received unexpected end of stream while receiving replication batches. " +
-                              "This might indicate an issue with network.", e);
-                throw;
-            }
-            catch (Exception e)
-            {
-                //if we are disposing, ignore errors
-                if (_cts.IsCancellationRequested)
-                    return;
-
-                DynamicJsonValue returnValue;
-
-                if (e.ExtractSingleInnerException() is MissingAttachmentException mae)
-                {
-                    returnValue = new DynamicJsonValue
-                    {
-                        [nameof(ReplicationMessageReply.Type)] = ReplicationMessageReply.ReplyType.MissingAttachments.ToString(),
-                        [nameof(ReplicationMessageReply.MessageType)] = messageType,
-                        [nameof(ReplicationMessageReply.LastEtagAccepted)] = -1,
-                        [nameof(ReplicationMessageReply.Exception)] = mae.ToString()
-                    };
-
-                    context.Write(writer, returnValue);
-                    writer.Flush();
-
-                    return;
-                }
-
-                if (_log.IsInfoEnabled)
-                    _log.Info($"Failed replicating documents {FromToString}.", e);
-
-                //return negative ack
-                returnValue = new DynamicJsonValue
-                {
-                    [nameof(ReplicationMessageReply.Type)] = ReplicationMessageReply.ReplyType.Error.ToString(),
-                    [nameof(ReplicationMessageReply.MessageType)] = messageType,
-                    [nameof(ReplicationMessageReply.LastEtagAccepted)] = -1,
-                    [nameof(ReplicationMessageReply.Exception)] = e.ToString()
-                };
-
-                context.Write(writer, returnValue);
-                writer.Flush();
-
-                throw;
-            }
         }
 
-        private void ReceiveSingleDocumentsBatch(TransactionOperationContext context, int replicatedItemsCount, int attachmentStreamCount)
+        protected override void OnDocumentsReceived()
+        {
+        }
+
+        protected override void OnAttachmentStreamsReceives(int attachmentStreamCount)
+        {
+        }
+
+        protected override void ReceiveSingleDocumentsBatch(TransactionOperationContext context, int replicatedItemsCount, int attachmentStreamCount, long lastEtag, IncomingReplicationStatsScope stats)
         {
             var sw = Stopwatch.StartNew();
-            var stats = new IncomingReplicationStatsScope(new IncomingReplicationRunStats());
 
             try
             {
                 using (_attachmentStreamsTempFile.Scope())
                 using (var shardAllocator = new IncomingReplicationAllocator(context.Allocator, maxSizeToSend: null))
+                using (var dataForReplicationCommand = new IncomingReplicationHandler.DataForReplicationCommand
                 {
-                    using (var networkStats = stats.For(ReplicationOperation.Incoming.Network))
+                    SupportedFeatures = SupportedFeatures,
+                    Logger = Logger
+                })
+                {
+                    try
                     {
-                        var reader = new Reader(_stream, _copiedBuffer, shardAllocator);
-
-                        var dictionary = ReadItemsFromSource(replicatedItemsCount, context, reader, stats);
-                        ReadAttachmentStreamsFromSource(attachmentStreamCount, context, reader, networkStats);
-
-                        foreach (var kvp in dictionary)
+                        using (var networkStats = stats.For(ReplicationOperation.Incoming.Network))
                         {
-                            var shard = kvp.Key;
-                            _replicationQueue.Items[shard].Add(kvp.Value);
-                        }
+                            var reader = new Reader(_stream, _copiedBuffer, shardAllocator);
 
-                        _replicationQueue.SendToShardCompletion.Wait();
-                        _replicationQueue.SendToShardCompletion = new CountdownEvent(_parent.Context.ShardCount);
+                            ReadItemsFromSource(replicatedItemsCount, context, context.Allocator, dataForReplicationCommand, reader, networkStats);
+                            ReadAttachmentStreamsFromSource(attachmentStreamCount, context, context.Allocator, dataForReplicationCommand, reader, networkStats);
+                            PrepareReplicationDataForShards(context, dataForReplicationCommand);
+                            
+                            _replicationQueue.SendToShardCompletion.Wait();
+                            _replicationQueue.SendToShardCompletion = new CountdownEvent(_parent.Context.ShardCount);
+                        }
+                    }
+                    catch (Exception e)
+                    {
+                        if (Logger.IsInfoEnabled)
+                        {
+                            //This is the case where we had a missing attachment, it is rare but expected.
+                            if (e.ExtractSingleInnerException() is MissingAttachmentException mae)
+                            {
+                                Logger.Info("Replication batch contained missing attachments will request the batch to be re-sent with those attachments.", mae);
+                            }
+                            else
+                            {
+                                Logger.Info("Failed to receive documents replication batch. This is not supposed to happen, and is likely a bug.", e);
+                            }
+                        }
+                        throw;
                     }
                 }
 
@@ -266,19 +108,14 @@ namespace Raven.Server.Documents.Sharding.Handlers
             }
         }
 
-        private Dictionary<int, List<ReplicationBatchItem>> ReadItemsFromSource(int replicatedDocs, TransactionOperationContext context, Reader reader, IncomingReplicationStatsScope stats)
+        private void PrepareReplicationDataForShards(TransactionOperationContext context, IncomingReplicationHandler.DataForReplicationCommand dataForReplicationCommand)
         {
-            var dict = new Dictionary<int, List<ReplicationBatchItem>>();
-
+            var dictionary = new Dictionary<int, List<ReplicationBatchItem>>();
             for (var shard = 0; shard < _parent.Context.ShardCount; shard++)
-            {
-                dict[shard] = new List<ReplicationBatchItem>();
-            }
+                dictionary[shard] = new List<ReplicationBatchItem>();
 
-            for (var i = 0; i < replicatedDocs; i++)
+            foreach (var item in dataForReplicationCommand.ReplicatedItems)
             {
-                var item = ReadItemFromSource(reader, context, context.Allocator, stats);
-
                 int shard = GetShardNumberForReplicationItem(context, item);
 
                 if (item is AttachmentReplicationItem attachment)
@@ -289,49 +126,41 @@ namespace Raven.Server.Documents.Sharding.Handlers
                         shardAttachments[attachment.Base64Hash] = attachment;
                 }
 
-                var list = dict[shard];
-                list.Add(item);
+                dictionary[shard].Add(item);
             }
 
-            return dict;
-        }
-
-        private void ReadAttachmentStreamsFromSource(int attachmentStreamCount, TransactionOperationContext context, Reader reader, IncomingReplicationStatsScope stats)
-        {
-            if (attachmentStreamCount == 0)
-                return;
-
-            for (var i = 0; i < attachmentStreamCount; i++)
+            if (dataForReplicationCommand.ReplicatedAttachmentStreams != null)
             {
-                var attachment = (AttachmentReplicationItem)ReplicationBatchItem.ReadTypeAndInstantiate(reader);
-                Debug.Assert(attachment.Type == ReplicationBatchItem.ReplicationItemType.AttachmentStream);
-
-                using (stats.For(ReplicationOperation.Incoming.AttachmentRead))
+                foreach (var attachment in dataForReplicationCommand.ReplicatedAttachmentStreams)
                 {
-                    attachment.ReadStream(context.Allocator, _attachmentStreamsTempFile);
-
                     for (var shard = 0; shard < _parent.Context.ShardCount; shard++)
                     {
                         var shardAttachments = _replicationQueue.AttachmentsPerShard[shard];
 
-                        if (shardAttachments.ContainsKey(attachment.Base64Hash))
+                        if (shardAttachments.ContainsKey(attachment.Key))
                         {
                             var attachmentStream = new AttachmentReplicationItem
                             {
                                 Type = ReplicationBatchItem.ReplicationItemType.AttachmentStream,
-                                Base64Hash = attachment.Base64Hash,
+                                Base64Hash = attachment.Key,
                                 Stream = new MemoryStream()
                             };
 
                             _attachmentStreamsTempFile._file.InnerStream.CopyTo(attachmentStream.Stream);
-                            shardAttachments[attachment.Base64Hash] = attachmentStream;
+                            shardAttachments[attachment.Key] = attachmentStream;
                         }
                     }
                 }
             }
+
+            foreach (var kvp in dictionary)
+            {
+                var shard = kvp.Key;
+                _replicationQueue.Items[shard].TryAdd(kvp.Value);
+            }
         }
 
-        private void SendHeartbeatStatusToSource(JsonOperationContext context, BlittableJsonTextWriter writer, long lastDocumentEtag, string handledMessageType)
+        internal override void SendHeartbeatStatusToSource(TransactionOperationContext context, BlittableJsonTextWriter writer, long lastDocumentEtag, string handledMessageType)
         {
             var heartbeat = new DynamicJsonValue
             {
@@ -345,6 +174,12 @@ namespace Raven.Server.Documents.Sharding.Handlers
             context.Write(writer, heartbeat);
             writer.Flush();
         }
+
+        protected override void AddReplicationPulse(ReplicationPulseDirection direction, string exceptionMessage = null)
+        {
+        }
+
+        protected override int GetNextReplicationStatsId() => _parent.GetNextReplicationStatsId();
 
 
         private readonly DocumentInfoHelper _documentInfoHelper = new DocumentInfoHelper();
